Fill page with search results initially
With a big enough monitor, or me in portrait mode, the observer might
never fire if the footer is already visible after the initial render.

Fix that by continuously calling `renderBatch` *until* the footer is
out of sight.

The change

```
-        this.observer.unobserve(this.observed);
+        this.observer.disconnect();
```
is to really free the `IntersectionObserver` resources as we don't ever
re-observe anything.